+++ packages/decap-cms-widget-markdown/src/MarkdownControl/renderers.js
@@ -97,7 +97,7 @@ const StyledTable = styled.table`
 `;
 
 const StyledTd = styled.td`
-  border: 2px solid black;
+  border: 1px solid black;
   padding: 8px;
   text-align: left;
 `;